Split pkg/genericapiserver/api/request from ObjectMeta
